Extract --pex-repository resolving from resolver.
The resolver module is huge and the code for resolution from a PEX had
nothing in common except the result type which is now extracted as well.
